+++ extensions/mypy_extensions.py
@@ -7,4 +7,16 @@ Example usage:
 
 # NOTE: This module must support Python 2.7 in addition to Python 3.x
 
-# (TODO: Declare TypedDict and other extensions here)
+
+def TypedDict(typename, fields):
+    """TypedDict creates a dictionary type that expects all of its
+    instances to have a certain set of keys, with each key
+    associated with a value of a consistent type. This expectation
+    is not checked at runtime but is only enforced by typecheckers.
+    """
+    def new_dict(*args, **kwargs):
+        return dict(*args, **kwargs)
+
+    new_dict.__name__ = typename
+    new_dict.__supertype__ = dict
+    return new_dict
