--- mypy/typing.pyi
@@ -1,6 +1,0 @@
-from typing import Dict, Type, TypeVar
-
-_T = TypeVar('_T')
-
-
-def TypedDict(typename: str, fields: Dict[str, Type[_T]]) -> Type[dict]: ...
