+++ test-data/unit/semanal-typeddict.test
@@ -1,12 +1,12 @@
 -- Semantic analysis of typed dicts
 
 [case testCanDefineTypedDictType]
-from mypy.typing import TypedDict
+from mypy_extensions import TypedDict
 Point = TypedDict('Point', {'x': int, 'y': int})
 [builtins fixtures/dict.pyi]
 [out]
 MypyFile:1(
-  ImportFrom:1(mypy.typing, [TypedDict])
+  ImportFrom:1(mypy_extensions, [TypedDict])
   AssignmentStmt:2(
     NameExpr(Point* [__main__.Point])
     TypedDictExpr:2(Point)))
@@ -14,38 +14,38 @@ MypyFile:1(
 -- Errors
 
 [case testTypedDictWithTooFewArguments]
-from mypy.typing import TypedDict
+from mypy_extensions import TypedDict
 Point = TypedDict('Point')  # E: Too few arguments for TypedDict()
 [builtins fixtures/dict.pyi]
 
 [case testTypedDictWithTooManyArguments]
-from mypy.typing import TypedDict
+from mypy_extensions import TypedDict
 Point = TypedDict('Point', {'x': int, 'y': int}, dict)  # E: Too many arguments for TypedDict()
 [builtins fixtures/dict.pyi]
 
 [case testTypedDictWithInvalidName]
-from mypy.typing import TypedDict
+from mypy_extensions import TypedDict
 Point = TypedDict(dict, {'x': int, 'y': int})  # E: TypedDict() expects a string literal as the first argument
 [builtins fixtures/dict.pyi]
 
 [case testTypedDictWithInvalidItems]
-from mypy.typing import TypedDict
+from mypy_extensions import TypedDict
 Point = TypedDict('Point', {'x'})  # E: TypedDict() expects a dictionary literal as the second argument
 [builtins fixtures/dict.pyi]
 
 -- NOTE: The following code works at runtime but is not yet supported by mypy.
 --       Keyword arguments may potentially be supported in the future.
 [case testTypedDictWithNonpositionalArgs]
-from mypy.typing import TypedDict
+from mypy_extensions import TypedDict
 Point = TypedDict(typename='Point', fields={'x': int, 'y': int})  # E: Unexpected arguments to TypedDict()
 [builtins fixtures/dict.pyi]
 
 [case testTypedDictWithInvalidItemName]
-from mypy.typing import TypedDict
+from mypy_extensions import TypedDict
 Point = TypedDict('Point', {int: int, int: int})  # E: Invalid TypedDict() field name
 [builtins fixtures/dict.pyi]
 
 [case testTypedDictWithInvalidItemType]
-from mypy.typing import TypedDict
+from mypy_extensions import TypedDict
 Point = TypedDict('Point', {'x': 1, 'y': 1})  # E: Invalid field type
 [builtins fixtures/dict.pyi]
